Add state files
git-svn-id: http://svn.macosforge.org/repository/odcctools/trunk@7 a073acdd-5acc-438f-8c47-dd119020c213

+++ Makefile
@@ -10,12 +10,14 @@ PATCHFILES=as/driver.c ld-Bstatic.diff
 
 ADDEDFILESDIR=$(TOPSRCDIR)/files
 ADDEDFILES=configure.ac Makefile.in include/config.h.in install-sh	\
-	config.guess config.sub
+	config.guess config.sub as/Makefile.in as/Makefile.ppc.in	\
+	as/Makefile.ppc64.in as/Makefile.i386.in as/Makefile.arch.in
 
 default: none
 
 clean:
 	rm -rf $(DISTDIR)
+	rm -rf .state.*
 
 none:
 	@echo "Please choose an action:"
@@ -25,22 +27,32 @@ none:
 
 
 extract:
-	if [ \! -d $(DISTDIR) ]; then				\
-		tar zxf $(DISTFILE);				\
+	if [ \! -f .state.extract ]; then			\
+		if [ \! -d $(DISTDIR) ]; then			\
+			tar zxf $(DISTFILE);			\
+		fi;						\
+		touch .state.extract;				\
 	fi
 
-patch: extract
-	for p in $(PATCHFILES); do				\
-		echo Applying patch $$p;			\
-		dir=`dirname $$p`;				\
-		( cd $(DISTDIR)/$$dir; 				\
-		  patch --posix -p0 < $(PATCHFILESDIR)/$$p );	\
-	done
-	for p in $(ADDEDFILES); do				\
-		echo Adding file $$p;				\
-		cp $(ADDEDFILESDIR)/$$p $(DISTDIR)/$$p;		\
-	done
+patch: extract 
+	if [ \! -f .state.patch ]; then				\
+		for p in $(PATCHFILES); do			\
+			echo Applying patch $$p;		\
+			dir=`dirname $$p`;			\
+			( cd $(DISTDIR)/$$dir; 			\
+			  patch --posix -p0 < $(PATCHFILESDIR)/$$p );	\
+		done;						\
+		for p in $(ADDEDFILES); do			\
+			echo Adding file $$p;			\
+			cp $(ADDEDFILESDIR)/$$p $(DISTDIR)/$$p;	\
+		done;						\
+		touch .state.patch;				\
+	fi
 
-regen:
-	cd $(DISTDIR) && autoheader;
-	cd $(DISTDIR) && autoconf;
+regen: patch
+	if [ \! -f .state.regen ]; then				\
+		( cd $(DISTDIR) &&				\
+		  autoheader &&					\
+		  autoconf );					\
+		touch .state.regen;				\
+	fi
